pyproject.toml
[tool.poetry]
name = "prismtoolbox"
version = "0.0.1"
description = "Toolbox for histopathology analysis"
authors = ["Loïc Le Bescond <loic.le.bescond@centralesupelec.fr>"]
packages = [
    { include = "prismtoolbox", from = "src"},
]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
openslide-python = "^1.3.1"
tiffslide = "^2.3.0"
opencv-python = "^4.8.1.78"
scipy = "^1.12.0"
h5py = "^3.10.0"
shapely = "^2.0.3"
matplotlib = "^3.8.3"
uuid = "^1.30"
pandas = "^2.2.1"

torch = { version = "^2.2.2", optional = true }
torchvision = { version = "^0.17.2", optional = true }
tqdm = { version = "^4.66.2", optional = true }
transformers = { version = "^4.38.2", optional = true }
timm = { version = "^0.9.16", optional = true }
scikit-learn = "^1.4.1.post1"
huggingface-hub = { version = "^0.22.1", optional = true }
umap-learn = {extras = ["plot"], version = "^0.5.5"}
cellpose = {version = "^3.0.7", optional = true}
dask = {extras = ["dataframe"], version = "^2024.3.1", optional = true}
ipywidgets = { version = "^8.1.2", optional = true }

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
isort = "^5.10.1"
pytest = "^7.1.3"
mkdocs-material = "^8.5.6"
mkdocstrings = {extras = ["python"], version = "^0.24.3"}
setuptools = "^69.2.0"
mkdocs-gen-files = "^0.5.0"
mkdocs-literate-nav = "^0.6.1"
mkdocs-section-index = "^0.3.8"
mkdocs-autorefs = "^1.0.1"
pyment = "^0.3.3"
pytest-subtests = "^0.12.1"

[tool.poetry.extras]
emb = ["torch", "torchvision", "tqdm", "transformers", "timm", "scikit-learn", "huggingface-hub", "umap-learn", "dask"]
seg = ["torch",  "torchvision", "cellpose", "ipywidgets"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.black]
line-length = 90
include = '\.pyi?$'
exclude = '''
/(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
'''

[tool.isort]
profile = "black"
skip_glob = ["*/__init__.py"]
